renaming some packages and files
