+++ akva.yaml
@@ -2,17 +2,9 @@ sinks:
   - type: cert
     path: /etc/nginx/certs/foo.cert
     frequency: 5s
-    vaultName: cjohnson-kv-test
+    vaultName: https://cmmkeyvaulttest.vault.azure.net/
     name: cjohnson-test-cert
     postChange: service nginx restart
     preChange: who knows
-    version: latest # or specific version number
-  - type: cert
-    path: /etc/nginx/certs/bar.cert
-    frequency: 5s
-    vaultName: cjohnson-kv-test
-    name: cjohnson-test-cert-2
-    postChange: service nginx restart
-    preChange: who knows
-    version: latest # or specific version number
+    version: latest
